First semi modular version.
