+++ scripts/pytest-vnet
@@ -1,0 +1,30 @@
+#!/usr/bin/env python3
+
+import sys
+import docker
+import pytest_vnet
+
+from pathlib import Path
+
+from pytest_vnet import PYTHON_VERSION
+
+
+client = docker.from_env()
+
+# Check if base netvm image is built
+try:
+	client.images.get("pytest-vnet:netvm")
+
+except docker.errors.ImageNotFound:
+	print("Couldn't find base netvm image, check build instructions on README.md")
+	sys.exit(1)
+
+# Check if python enabled netvm image is ready
+try:
+	client.images.get(f"pytest-vnet:netvm-{PYTHON_VERSION}")
+
+except docker.errors.ImageNotFound:
+	# Attempt installation
+	pytest_vnet.prepare_container()
+
+pytest_vnet.run_in_netvm(Path(sys.argv[1]))
